familygraph add gradle and update api
