+++ familygraph/gradle.properties
@@ -1,0 +1,7 @@
+VERSION_NAME=0.1-SNAPSHOT
+VERSION_CODE=0.1
+GROUP=com.familygraph
+
+POM_NAME=FamilyGraph
+POM_ARTIFACT_ID=library
+POM_PACKAGING=aar
